Refactor motion handler
