+++ test/org/traccar/handler/MotionHandlerTest.java
@@ -1,4 +1,4 @@
-package org.traccar;
+package org.traccar.handler;
 
 import static org.junit.Assert.assertEquals;
 
@@ -8,7 +8,7 @@ import org.traccar.model.Position;
 public class MotionHandlerTest {
 
     @Test
-    public void testCalculateMotion() throws Exception {
+    public void testCalculateMotion() {
 
         MotionHandler motionHandler = new MotionHandler(0.01);
